moves types common to clearing house documents to the common types folder
