+++ app/types/common/documents.ts
@@ -1,0 +1,23 @@
+type SubDocument = {
+  identifier: string
+}
+
+export type ELink = {
+  url: string
+}
+
+export type DocumentRequest = {
+  header: {
+    identifier: string
+  }
+}
+
+export type KeywordType = {
+  name: string
+  identifier: string
+}
+
+export type Keywords = {
+  processedKeywords: Array<SubDocument>
+  otherKeywords: string
+}
